Implement zope.implements for Python2 and Python3
