+++ CHANGES.txt
@@ -6,9 +6,9 @@ Changes
 
 - compatibility for python 3
 - update code to work with python2 and python3
-- update doctests to work with python2 and python3
 - adjust egg to work with newer version of setuptools
-- reimplment ``zope.interface.implements`` with a grokker
+- reimplement ``zope.interface.implements`` with a grokker
+  for both python2 and python3
 
 2.5 (2012-05-01)
 ----------------
