+++ src/grokcore/component/__init__.py
@@ -19,8 +19,6 @@ adapts.__doc__ = "Declares the types of objects that a multi-adapter adapts."
 
 from zope.interface import implementsOnly, classProvides
 
-from grokcore.component.compat import implements
-
 from martian import baseclass
 from martian.error import GrokError, GrokImportError
 from martian import ClassGrokker, InstanceGrokker, GlobalGrokker
@@ -45,6 +43,7 @@ from grokcore.component.directive import (
     path,
     provides,
     title,
+    implements,
     )
 
 from grokcore.component.decorators import (
